app/requirements.txt
Fiona==1.8.6
geopandas==0.5.1
numpy
pandas
Pillow==7.1.2
pyodbc
pyproj==1.9.6
Shapely==1.7.0
SQLAlchemy
streamlit
boka_tools-0.0.4-py3-none-any.whl